% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elt.R
\name{add_metadata}
\alias{add_metadata}
\title{Add Metadata Timestamps to a Data Frame}
\usage{
add_metadata(df)
}
\arguments{
\item{df}{A data frame or tibble. The input data to which metadata will be added.}
}
\value{
A tibble with two additional columns:
}
\description{
Adds two metadata columns to a data frame: the current system time (\code{updated_at}) and a local-adjusted version (\code{local_updated_at}) offset by 6 hours.
}
