% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elt.R
\name{create_schema}
\alias{create_schema}
\title{Create a Database Schema if It Doesn't Exist}
\usage{
create_schema(con, schema_name)
}
\arguments{
\item{con}{A DBI database connection object. Typically created with \code{DBI::dbConnect()}.}

\item{schema_name}{A character string specifying the name of the schema to create.}
}
\value{
The number of rows affected (usually \code{0} or a backend-specific result), invisibly.
}
\description{
Executes a SQL command to create a schema in the connected database, if it does not already exist.
}
