% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tblx}
\alias{tblx}
\title{I just like to make it easier for myself}
\usage{
tblx(schema, table)
}
\arguments{
\item{schema}{Character}

\item{table}{Character}
}
\description{
I just like to make it easier for myself
}
