% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elt.R
\name{update_ncaa_db}
\alias{update_ncaa_db}
\title{Update WBB Database Tables from \code{wehoop} Data}
\usage{
update_ncaa_db(con, seasons, overwrite = FALSE)
}
\arguments{
\item{con}{A DBI database connection object.}

\item{seasons}{A vector of numeric or character values indicating the seasons to load.}

\item{overwrite}{A logical flag. If \code{TRUE}, existing records will be overwritten. If \code{FALSE}, only matching records will be deleted and new data appended.}
}
\value{
Invisibly returns \code{NULL}. Primarily called for its side effects on the database.
}
\description{
Loads and updates WBB schedule, player box score, and team box score data for the specified seasons
into a connected database using the \code{load_data()} utility.
}
\details{
This function performs the following steps:
\itemize{
\item Loads WBB schedule data (removing the \code{venue_capacity} column if present).
\item Loads WBB player box score data for the given seasons.
\item Loads WBB team box score data
}

All data is written to the \code{WBB} schema in the database:
\itemize{
\item \code{SCHEDULE} table for schedule data
\item \code{PLAYER_BOX_STG} for player box score data
\item \code{TEAM_BOX_STG} for team box score data
}
}
