% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elt.R
\name{update_ncaa_pbp}
\alias{update_ncaa_pbp}
\title{Update the wehoop database}
\usage{
update_ncaa_pbp(
  dbdir = "~/.db",
  dbname = "belle",
  force_rebuild = FALSE,
  dbpath = Sys.getenv("DB_PATH_WNBA")
)
}
\arguments{
\item{dbdir}{Character. The directory where the database is stored. Defaults to \verb{~/.db}.}

\item{dbname}{Character. DB name, defaults to my dog \code{belle}}

\item{force_rebuild}{Logical. If \code{TRUE}, forces a rebuild of the database. Defaults to \code{FALSE}.}

\item{dbpath}{Path for custom db}
}
\value{
No return value. Updates the nflfastR database in place.
}
\description{
This function is a wrapper that updates the \code{wehoop} database by
calling \code{wehoop::update_wbb_db()}.
}
\examples{
\dontrun{
update_db()  # Updates the database in the default location
}
}
