% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{update_wnba_db}
\alias{update_wnba_db}
\title{main function to update tables that are related to \verb{\{wehoop\} WNBA data}.}
\usage{
update_wnba_db(seasons = NULL)
}
\arguments{
\item{seasons}{numeric vector to indentify seasons to back filter for and backfill.
Note: this should typically only be used with \code{wehoop::most_recent_season()} or maintain
the default value of \code{NULL}, as datasets have different years that started collecting. It is
easier to do a full refresh than handle all the different conditions.}
}
\description{
This function will update tables based on the ENV variable \code{DB_PATH} that is used in
\code{load_data()}.
}
