% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{load_data}
\alias{load_data}
\title{Update the specified database table}
\usage{
load_data(
  df,
  table_name,
  schema_name = "BASE",
  db_path = Sys.getenv("DB_PATH_WNBA"),
  seasons = NULL
)
}
\arguments{
\item{df}{A data frame to load into the db}

\item{table_name}{name of table to use in db}

\item{schema_name}{name of schema to use, defaults to `"BASE"``}

\item{db_path}{location of db}

\item{seasons}{seasons to update when using existing database.}
}
\value{
No return value. Updates the specified db
}
\description{
this function will load data into the db. It can handle creating new schemas
and tables, and will overwrite data if there is a specified season or set of seasons.
}
